# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type BoardSummaryRole = typeof BoardSummaryRole[keyof typeof BoardSummaryRole];


export const BoardSummaryRole = {
  owner: 'owner',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface BoardSummary {
  id: string;
  name: string;
  ownerId: string;
  /** @nullable */
  ownerName: string | null;
  role: BoardSummaryRole;
  collaboratorCount?: number;
  createdAt: string;
  updatedAt: string;
}

export type BoardRole = typeof BoardRole[keyof typeof BoardRole];


export const BoardRole = {
  owner: 'owner',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface Board {
  id: string;
  name: string;
  ownerId: string;
  /** @nullable */
  ownerName: string | null;
  role: BoardRole;
  /** Serialized canvas state (array of drawn element objects), stored as opaque JSON. */
  elements: unknown[];
  createdAt: string;
  updatedAt: string;
}

export interface BoardInput {
  /** @minLength 1 */
  name: string;
}

export interface BoardUpdate {
  /** @minLength 1 */
  name?: string;
  elements?: unknown[];
}

export type CollaboratorRole = typeof CollaboratorRole[keyof typeof CollaboratorRole];


export const CollaboratorRole = {
  owner: 'owner',
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface Collaborator {
  userId: string;
  boardId: string;
  role: CollaboratorRole;
  /** @nullable */
  name: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export type CollaboratorInputRole = typeof CollaboratorInputRole[keyof typeof CollaboratorInputRole];


export const CollaboratorInputRole = {
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface CollaboratorInput {
  userId: string;
  role: CollaboratorInputRole;
}

export type CollaboratorRoleUpdateRole = typeof CollaboratorRoleUpdateRole[keyof typeof CollaboratorRoleUpdateRole];


export const CollaboratorRoleUpdateRole = {
  editor: 'editor',
  viewer: 'viewer',
} as const;

export interface CollaboratorRoleUpdate {
  role: CollaboratorRoleUpdateRole;
}

export interface ChatMessage {
  id: string;
  boardId: string;
  userId: string;
  /** @nullable */
  userName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  message: string;
  createdAt: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type SearchUsersParams = {
q: string;
};