# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: boards
    description: Whiteboard board management
  - name: collaborators
    description: Board collaborator management
  - name: chat
    description: Board chat messages
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /boards:
    get:
      operationId: listBoards
      tags: [boards]
      summary: List boards owned by or shared with the current user
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BoardSummary"
    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a new board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardInput"
      responses:
        "201":
          description: Board created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"

  /boards/{id}:
    get:
      operationId: getBoard
      tags: [boards]
      summary: Get a board and its saved canvas state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Board detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "404":
          description: Board not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateBoard
      tags: [boards]
      summary: Rename a board or update its saved canvas state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardUpdate"
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "404":
          description: Board not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteBoard
      tags: [boards]
      summary: Delete a board (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Board not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /boards/{id}/collaborators:
    get:
      operationId: listCollaborators
      tags: [collaborators]
      summary: List collaborators on a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of collaborators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collaborator"
    post:
      operationId: addCollaborator
      tags: [collaborators]
      summary: Invite a collaborator to a board by their user id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaboratorInput"
      responses:
        "201":
          description: Collaborator added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
        "404":
          description: Board or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /boards/{id}/collaborators/{userId}:
    patch:
      operationId: updateCollaboratorRole
      tags: [collaborators]
      summary: Update a collaborator's permission level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaboratorRoleUpdate"
      responses:
        "200":
          description: Updated collaborator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: removeCollaborator
      tags: [collaborators]
      summary: Remove a collaborator from a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /boards/{id}/chat:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: Get recent chat history for a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  /users/search:
    get:
      operationId: searchUsers
      tags: [collaborators]
      summary: Search users by name or email to invite as collaborators
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthUser"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    BoardSummary:
      type: object
      required: [id, name, ownerId, ownerName, role, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        ownerName:
          type: [string, 'null']
        role:
          type: string
          enum: [owner, editor, viewer]
        collaboratorCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Board:
      type: object
      required: [id, name, ownerId, ownerName, role, elements, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        ownerName:
          type: [string, 'null']
        role:
          type: string
          enum: [owner, editor, viewer]
        elements:
          description: Serialized canvas state (array of drawn element objects), stored as opaque JSON.
          type: array
          items: {}
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BoardInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    BoardUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        elements:
          type: array
          items: {}

    Collaborator:
      type: object
      required: [userId, boardId, role, name, email, profileImageUrl]
      properties:
        userId:
          type: string
        boardId:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]
        name:
          type: [string, 'null']
        email:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']

    CollaboratorInput:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [editor, viewer]

    CollaboratorRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [editor, viewer]

    ChatMessage:
      type: object
      required: [id, boardId, userId, userName, profileImageUrl, message, createdAt]
      properties:
        id:
          type: string
        boardId:
          type: string
        userId:
          type: string
        userName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        message:
          type: string
        createdAt:
          type: string
          format: date-time