# lib/api-zod/src/generated/types/board.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoardRole } from './boardRole';

export interface Board {
  id: string;
  name: string;
  ownerId: string;
  /** @nullable */
  ownerName: string | null;
  role: BoardRole;
  /** Serialized canvas state (array of drawn element objects), stored as opaque JSON. */
  elements: unknown[];
  createdAt: Date;
  updatedAt: Date;
}