# lib/api-zod/src/generated/types/boardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BoardSummaryRole } from './boardSummaryRole';

export interface BoardSummary {
  id: string;
  name: string;
  ownerId: string;
  /** @nullable */
  ownerName: string | null;
  role: BoardSummaryRole;
  collaboratorCount?: number;
  createdAt: Date;
  updatedAt: Date;
}