# lib/api-zod/src/generated/types/boardUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BoardUpdate {
  /** @minLength 1 */
  name?: string;
  elements?: unknown[];
}