# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: string;
  boardId: string;
  userId: string;
  /** @nullable */
  userName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  message: string;
  createdAt: Date;
}