# lib/api-zod/src/generated/types/collaborator.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CollaboratorRole } from './collaboratorRole';

export interface Collaborator {
  userId: string;
  boardId: string;
  role: CollaboratorRole;
  /** @nullable */
  name: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}