# lib/api-zod/src/generated/types/collaboratorRoleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CollaboratorRoleUpdateRole } from './collaboratorRoleUpdateRole';

export interface CollaboratorRoleUpdate {
  role: CollaboratorRoleUpdateRole;
}